youtube/requirements.txt
amqp==5.0.5
billiard==3.6.3.0
cachetools==4.2.1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Flask==1.1.2
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
googleapis-common-protos==1.53.0
h11==0.12.0
httplib2==0.19.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
kombu==5.0.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
nltk==3.5
packaging==20.9
prompt-toolkit==3.0.16
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.1
pyparsing==2.4.7
pytz==2021.1
rake-nltk==1.0.4
redis==3.5.3
regex==2020.11.13
requests==2.25.1
rsa==4.7.2
six==1.15.0
starlette==0.13.6
tqdm==4.58.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
uvicorn==0.13.4
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1